Add support for enforcing angular.mock.module
This is needed when migrating angular.js to use webpack, and the existing unit tests call `module('myModule')`. This causes problems as `module` is a reserved/special word for webpack use, and the unit tests won't compile. The solution is to use `angular.mock.module('myModule')`. This rule change adds a `webpack-support` option (open to renaming this), and adds a fixer that can sort out the problems found (to save on Ctrl+F'ing through the code and finding issues).